Move flank/wrapper methods to new module to use in tul_cob.
