+++ lib/cob_index/macros.rb
@@ -3,6 +3,7 @@
 module CobIndex::Macros
   autoload :Booster, "cob_index/macros/booster"
   autoload :Custom, "cob_index/macros/custom"
+  autoload :Wrapper, "cob_index/macros/wrapper"
   autoload :Marc21, "cob_index/macros/marc21"
   autoload :MarcFormats, "cob_index/macros/marc_format_classifier"
 end
