+++ lib/cob_index/macros/custom.rb
@@ -636,25 +636,6 @@ module CobIndex::Macros::Custom
     end
   end
 
-  def wrap_begin_end(*args)
-    lambda do |record, accumulator, context|
-      accumulator.map! { |v| flank v }
-    end
-  end
-
-  def flank(string = "", starts = nil, ends = nil)
-    starts ||= "matchbeginswith"
-    ends ||= "matchendswith"
-    if !string.to_s.empty? && !string.match(/^#{starts}/)
-
-      first_word = string.split.first
-      last_word = string.split.last
-      "#{starts}#{first_word} #{string} #{ends}#{last_word}"
-    else
-      string
-    end
-  end
-
   def suppress_items
     lambda do |rec, acc, context|
       full_text_link = rec.fields("856").select { |field| field["u"] }
