+++ spec/cob_index/macros/custom_spec.rb
@@ -54,37 +54,6 @@ RSpec.describe "custom methods" do
     end
   end
 
-  describe "#flank(field)" do
-    let(:input) {}
-    subject { CobIndex::Macros::Custom.flank input }
-    context "nil" do
-      it "returns an empty string" do
-        expect(subject).to be_nil
-      end
-    end
-
-    context "empty string" do
-      let(:input) { "" }
-      it "returns an empty string" do
-        expect(subject).to eq("")
-      end
-    end
-
-    context "non empty string" do
-      let(:input) { "foo bar buzz" }
-      it "returns a flanked string" do
-        expect(subject).to eq("matchbeginswithfoo foo bar buzz matchendswithbuzz")
-      end
-    end
-
-    context "a string that is flanked" do
-      let(:input) { "matchbeginswithfoo foo bar buzz matchendswithbar" }
-      it "does not reflank a string" do
-        expect(subject).to eq(input)
-      end
-    end
-  end
-
   describe "#creator_name_trim_punctuation(name)" do
     context "removes trailing comma, slash" do
       let(:input) { "Richard M. Restak." }
